// codegen/cms/pages/models/AttachToLangPrimaryRequestVNext.ts
/**
 * Pages
 * Use these endpoints for interacting with Landing Pages and Site Pages
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
* Request body object fro attaching objects to multi-language groups.
*/
export class AttachToLangPrimaryRequestVNext {
    /**
    * Designated language of the object to add to a multi-language group.
    */
    'language': string;
    /**
    * ID of the object to add to a multi-language group.
    */
    'id': string;
    /**
    * ID of primary language object in multi-language group.
    */
    'primaryId': string;
    /**
    * Primary language of the multi-language group.
    */
    'primaryLanguage'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "language",
            "baseName": "language",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "primaryId",
            "baseName": "primaryId",
            "type": "string",
            "format": ""
        },
        {
            "name": "primaryLanguage",
            "baseName": "primaryLanguage",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AttachToLangPrimaryRequestVNext.attributeTypeMap;
    }

    public constructor() {
    }
}