// codegen/cms/pages/models/BatchInputContentFolder.ts
/**
 * Pages
 * Use these endpoints for interacting with Landing Pages and Site Pages
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ContentFolder } from '../models/ContentFolder';

/**
* Wrapper for providing an array of content folders as inputs.
*/
export class BatchInputContentFolder {
    /**
    * Content folders to input.
    */
    'inputs': Array<ContentFolder>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<ContentFolder>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BatchInputContentFolder.attributeTypeMap;
    }

    public constructor() {
    }
}