// codegen/cms/pages/models/BatchInputJsonNode.ts
/**
 * Pages
 * Use these endpoints for interacting with Landing Pages and Site Pages
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
* Wrapper for providing an array of JSON nodes as inputs.
*/
export class BatchInputJsonNode {
    /**
    * JSON nodes to input.
    */
    'inputs': Array<any>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<any>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BatchInputJsonNode.attributeTypeMap;
    }

    public constructor() {
    }
}